Step17: 카프카 설정 변경 및 메시지 Producer
