+++ src/main/java/io/hhplus/ecommerce/config/kafka/KafkaTopicConfig.java
@@ -9,7 +9,7 @@ import org.springframework.kafka.config.TopicBuilder;
 public class KafkaTopicConfig {
 
     @Bean
-    public NewTopic paymentSuccessTopic() {
-        return TopicBuilder.name("payment-success").build();
+    public NewTopic paymentCompleteTopic() {
+        return TopicBuilder.name("payment-complete").build();
     }
 }
