--- src/main/java/io/hhplus/ecommerce/payment/application/eventListener/PaymentEventHandler.java
@@ -1,25 +1,0 @@
-package io.hhplus.ecommerce.payment.application.eventListener;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.hhplus.ecommerce.payment.domain.event.PaymentCompleteEvent;
-import lombok.RequiredArgsConstructor;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.event.TransactionPhase;
-import org.springframework.transaction.event.TransactionalEventListener;
-
-@Component
-@RequiredArgsConstructor
-public class PaymentEventHandler {
-
-    private final KafkaTemplate<String, String> kafkaTemplate;
-
-    @Async
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public void paymentCompleteEventHandler(PaymentCompleteEvent event) throws JsonProcessingException {
-        ObjectMapper objectMapper = new ObjectMapper();
-        kafkaTemplate.send(event.getTopic(), objectMapper.writeValueAsString(event));
-    }
-}
