+++ src/main/java/io/hhplus/ecommerce/payment/domain/event/PaymentCompleteEvent.java
@@ -7,13 +7,13 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PaymentCompleteEvent {
-    private String topic;
+
     private Long userSeq;
+
     private Long orderId;
 
     public static PaymentCompleteEvent of(Long userSeq, Long orderId) {
         return PaymentCompleteEvent.builder()
-                .topic("payment-success")
                 .userSeq(userSeq)
                 .orderId(orderId)
                 .build();
